Убрала лишний стор
